+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 3.0.0-nullsafety.0
 
 * Migrated to null safety.
+* Fixed an issue with usage of double.infinity in DeltaIterator.
+  Replaced with special value of DeltaIterator.maxLength.
 
 ## 2.0.1